requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
click==8.0.4
colorama==0.4.4
databases==0.5.5
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
joblib==1.1.0
Mako==1.2.0
MarkupSafe==2.1.1
networkx==2.7.1
nltk==3.7
numpy==1.22.3
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
regex==2022.3.15
scipy==1.8.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tqdm==4.63.1
typing-extensions==4.1.1
uvicorn==0.17.6
